requirements.txt
sqlalchemy~=2.0.44
fastapi~=0.120.0
pydantic~=2.12.3
uvicorn~=0.38.0
python-dotenv~=1.1.1
psycopg[binary]~=3.2.12

